+++ noodles-vcf/src/io/writer/header/record/value/map/format.rs
@@ -7,12 +7,12 @@ use self::{number::write_number, ty::write_type};
 use super::{
     write_delimiter, write_description_field, write_key, write_other_fields, write_separator,
 };
-use crate::header::{
-    record::value::{
-        map::{format::Type, Format},
-        Map,
+use crate::header::record::value::{
+    map::{
+        format::{Number, Type},
+        Format,
     },
-    Number,
+    Map,
 };
 
 pub(crate) fn write_format<W>(writer: &mut W, format: &Map<Format>) -> io::Result<()>